feat(bet4-lifecycle): extract build-lifecycle + change tools into scoped ToolPacks (EP-8DC217EB BET-4)
Process-Spine-Decision: BET-4 drain-to-zero (plan §4) — behavior-identical; imports shared build helpers; only principle_decide remains inline after this.
BI: BI-2B7EE073
