+++ apps/web/lib/mcp-tools.ts
@@ -566,69 +566,6 @@ export const PLATFORM_TOOLS: ToolDefinition[] = [
     sideEffect: false,
     buildPhases: ["review", "ship"],
   },
-  {
-    name: "update_feature_brief",
-    description: "Save the Feature Brief for the current build. Build ID is auto-resolved.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        title: { type: "string", description: "Feature title" },
-        description: { type: "string", description: "Plain-language feature description" },
-        portfolioContext: { type: "string", description: "Portfolio slug that owns this feature. For internal platform/meta work (a change to Build Studio, the portal, or platform tooling itself) pass an empty string rather than forcing a customer-facing portfolio." },
-        targetRoles: { type: "array", items: { type: "string" }, description: "Roles that will use this feature. For internal platform/meta work use internal operator roles (e.g. platform operator, admin) — never customer." },
-        inputs: { type: "array", items: { type: "string" }, description: "User inputs the feature accepts" },
-        dataNeeds: { type: "string", description: "What data the feature stores" },
-        acceptanceCriteria: { type: "array", items: { type: "string" }, description: "What done looks like. Every requirement the user stated explicitly (exact formats, examples, behaviors) MUST appear as its own criterion, preserved faithfully — never substitute a different format or behavior for one the user specified." },
-        fixContext: {
-          type: "object",
-          description: "For fix builds (kind=fix): the defect diagnosis. reproSteps, rootCause, and fixApproach are all required before a fix build can advance to plan. Merged into any existing fixContext, so partial updates accumulate.",
-          properties: {
-            reproSteps: { type: "string", description: "How to reproduce the defect" },
-            expected: { type: "string", description: "Expected behavior" },
-            actual: { type: "string", description: "Actual (buggy) behavior" },
-            rootCause: { type: "string", description: "Identified root cause (file/function + why)" },
-            fixApproach: { type: "string", description: "Smallest correct fix + regression test" },
-            severity: { type: "string" },
-          },
-        },
-      },
-      required: ["title", "description", "portfolioContext", "targetRoles", "dataNeeds", "acceptanceCriteria"],
-    },
-    requiredCapability: "view_platform",
-    executionMode: "immediate",
-    sideEffect: true,
-    buildPhases: ["ideate"],
-  },
-  {
-    name: "register_digital_product_from_build",
-    description: "Register or update a DigitalProduct from the current build. Build ID is auto-resolved. Requires approval.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        name: { type: "string", description: "Product name" },
-        portfolioSlug: { type: "string", description: "Portfolio slug to assign to" },
-      },
-      required: ["name", "portfolioSlug"],
-    },
-    requiredCapability: "manage_capabilities",
-    sideEffect: true,
-    buildPhases: ["ship"],
-  },
-  {
-    name: "create_build_epic",
-    description: "Create an Epic and backlog items for a shipped build. All IDs are auto-resolved.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        title: { type: "string", description: "Epic title" },
-      },
-      required: ["title"],
-    },
-    requiredCapability: "manage_capabilities",
-    executionMode: "immediate",
-    sideEffect: true,
-    buildPhases: ["ship"],
-  },
   // ─── Build Notes Tool ───────────────────────────────────────────────────
   {
     name: "save_build_notes",
@@ -683,21 +620,6 @@ export const PLATFORM_TOOLS: ToolDefinition[] = [
     executionMode: "immediate",
     sideEffect: false,
     buildPhases: ["ideate", "plan", "build", "review", "ship"],
-    annotations: { readOnlyHint: true, idempotentHint: true },
-  },
-  {
-    name: "verification_preflight",
-    description: "Deterministic verify-phase preflight (EP-VERIFY-PROC). Returns MUST_ADVANCE (evidence already sufficient — do not re-test), BLOCKED (a prerequisite is missing — report it, do not fabricate a result), or CAN_TEST (proceed to functional verification). Call this BEFORE attempting verification so testability is a procedural verdict, not a judgment call.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        buildId: { type: "string", description: "Optional FB-* build ID. Omit to target the current active build." },
-      },
-    },
-    requiredCapability: "view_platform",
-    executionMode: "immediate",
-    sideEffect: false,
-    buildPhases: ["build", "review"],
     annotations: { readOnlyHint: true, idempotentHint: true },
   },
   {
@@ -933,20 +855,6 @@ export const PLATFORM_TOOLS: ToolDefinition[] = [
     buildPhases: ["ideate", "plan", "build", "review"],
   },
   {
-    name: "start_build",
-    description: "Initialize the build workspace. Call this ONCE at the start of the build phase. Verifies the sandbox container is running and creates a git branch for this build. If it returns 'not running', call diagnose_sandbox and use the returned recovery actions.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        buildId: { type: "string", description: "Optional FB-* build ID. Omit to target the current active build. Supply it to drive a specific build when several builds are in-flight (e.g. an autonomous batch)." },
-      },
-    },
-    requiredCapability: "view_platform",
-    executionMode: "immediate",
-    sideEffect: false,
-    buildPhases: ["build"],
-  },
-  {
     name: "run_sandbox_tests",
     description: "Run unit tests and typecheck inside the sandbox container. Set auto_fix to true to automatically diagnose and fix failures (up to 3 attempts).",
     inputSchema: {
@@ -1087,49 +995,6 @@ export const PLATFORM_TOOLS: ToolDefinition[] = [
     executionMode: "immediate",
     sideEffect: false, // sandbox-isolated; inner calls are read-only + governed per call
   },
-  {
-    name: "validate_schema",
-    description: "Validate the Prisma schema in the sandbox for common errors: missing inverse relations, undefined types, unindexed foreign keys. MUST be called before running prisma migrate. Returns specific errors with fix instructions.",
-    inputSchema: { type: "object", properties: {} },
-    requiredCapability: "view_platform",
-    executionMode: "immediate",
-    sideEffect: false,
-    buildPhases: ["build"],
-  },
-  {
-    name: "deploy_feature",
-    description: "Extract the git diff from sandbox and deploy to the platform.",
-    inputSchema: { type: "object", properties: {} },
-    requiredCapability: "manage_capabilities",
-    executionMode: "immediate",
-    sideEffect: true,
-    buildPhases: ["ship"],
-  },
-  {
-    name: "create_portal_pr",
-    description: "Create a pull request on the portal's own repository from the current build's diff. Runs pre-PR security gates (security scan, destructive ops, architecture compliance, dependency audit). If all gates pass and the build is fully verified, auto-merges via squash. If any gate fails, creates the PR with findings and requests human review.",
-    inputSchema: { type: "object", properties: {} },
-    requiredCapability: "manage_capabilities",
-    executionMode: "immediate",
-    sideEffect: true,
-    buildPhases: ["ship"],
-  },
-  {
-    name: "set_change_disposition",
-    description: "Record the human's final call on whether the current change is kept private (on the user's own system) or shared with the community. Use after presenting the Keep/Share suggestion at ship time. A change must be 'shareable' before contribute_to_hive or a public-hive PR will share it; the default is 'private' (fail-closed), so inaction never shares.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        disposition: { type: "string", enum: ["private", "shareable"], description: "'private' keeps the change on the user's system; 'shareable' clears it to be contributed to the community." },
-        reason: { type: "string", description: "Optional short note on why this disposition was chosen." },
-      },
-      required: ["disposition"],
-    },
-    requiredCapability: "manage_capabilities",
-    executionMode: "immediate",
-    sideEffect: true,
-    buildPhases: ["ship"],
-  },
   // ─── Email setup (PBI-INV-04 Phase 2) ──────────────────────────────────
   // Lets the onboarding/COO coworker walk a non-technical operator through
   // configuring their OWN outbound email (SMTP). Operator-only
@@ -1165,64 +1030,6 @@ export const PLATFORM_TOOLS: ToolDefinition[] = [
     autoApproveWhen: async () => {
       return false;
     },
-  },
-  {
-    name: "run_ux_test",
-    description: "Run natural-language UX test cases against the sandbox using AI-powered browser automation (browser-use). Each test case is a plain English assertion that the AI agent verifies by driving a real browser. Returns structured pass/fail results with screenshots. NOTE: UX verification runs automatically on review-phase entry via build/review.verify — this tool is retained for ad-hoc manual invocations only and is not in the review phase's tool allowlist.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        tests: {
-          type: "array",
-          items: { type: "string" },
-          description: "Natural-language test assertions. If omitted, auto-generates from acceptance criteria.",
-        },
-      },
-    },
-    requiredCapability: "view_platform",
-    executionMode: "immediate",
-    sideEffect: false,
-    // Empty buildPhases — the Inngest handler owns UX verification during
-    // review; leaving this in the review allowlist leads to duplicate runs.
-    // Kept callable outside a build phase for Dev-mode / debugging use.
-    buildPhases: [],
-  },
-  // ─── Codebase Access Tools ──────────────────────────────────────────────────
-  {
-    name: "start_ideate_research",
-    description: "Signal that you have enough context from the user to begin codebase research and draft the design document. Call this AFTER the user answers your questions (intent gate, reusability scope). The system will search the codebase, analyze patterns, and draft the design doc automatically. You do NOT need to call search_project_files or read_project_file yourself — this tool handles all research.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        reusabilityScope: { type: "string", enum: ["one_off", "parameterizable", "already_generic"], description: "The user's reusability preference" },
-        userContext: { type: "string", description: "Summary of what the user wants, including any answers to clarifying questions" },
-        buildId: { type: "string", description: "Optional FB-* build ID. Omit to target the current active build (the ambient ideate conversation). Supply it to drive a specific build's research when several builds are in-flight (e.g. an autonomous 20-build batch)." },
-      },
-      required: ["reusabilityScope", "userContext"],
-    },
-    requiredCapability: "view_platform",
-    executionMode: "immediate",
-    sideEffect: true,
-    buildPhases: ["ideate"],
-  },
-  {
-    name: "start_scout_research",
-    description: "Start a fast codebase scout + URL parse before asking clarification questions. Call this immediately after the user describes their feature. Returns immediately — results appear in Build Studio Context on the next turn.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        externalUrls: {
-          type: "array",
-          items: { type: "string" },
-          description: "Any URLs the user mentioned (website, design doc, reference). Will be fetched and parsed for domain structure.",
-        },
-      },
-      required: [],
-    },
-    requiredCapability: "view_platform",
-    executionMode: "immediate",
-    sideEffect: true,
-    buildPhases: ["ideate"],
   },
   // ─── Manifest Tools ────────────────────────────────────────────────────────
   {
@@ -2151,135 +1958,6 @@ export async function executeTool(
       };
     }
 
-    case "update_feature_brief": {
-      const buildId = await resolveActiveBuildId(userId, extractBuildIdHint(params));
-      if (!buildId) return { success: false, error: "No active build", message: "No active build found" };
-      const { updateFeatureBrief } = await import("@/lib/actions/build");
-      // Merge OVER the existing brief. updateFeatureBrief persists the whole
-      // brief object, so rebuilding it from scratch would clobber any field the
-      // caller omits — notably fixContext, which the fix-flow ideate phase fills
-      // incrementally. A fix build is also promoted with title/description
-      // pre-seeded; without this merge a fixContext-only update would blank them
-      // and fail validateFeatureBrief.
-      const prior = (await prisma.featureBuild.findUnique({
-        where: { buildId },
-        select: { brief: true },
-      }))?.brief as Partial<import("@/lib/feature-build-types").FeatureBrief> | null;
-      const str = (key: string, fallback: string) =>
-        params[key] != null ? String(params[key]) : fallback;
-      const arr = (key: string, fallback: string[]) =>
-        Array.isArray(params[key]) ? (params[key] as unknown[]).map(String) : fallback;
-      const incomingFix = params["fixContext"] && typeof params["fixContext"] === "object" && !Array.isArray(params["fixContext"])
-        ? (params["fixContext"] as Record<string, unknown>)
-        : null;
-      const mergedFix = (prior?.fixContext || incomingFix)
-        ? { ...(prior?.fixContext ?? {}), ...(incomingFix ?? {}) }
-        : undefined;
-      const brief = {
-        title: str("title", prior?.title ?? ""),
-        description: str("description", prior?.description ?? ""),
-        portfolioContext: str("portfolioContext", prior?.portfolioContext ?? ""),
-        targetRoles: arr("targetRoles", prior?.targetRoles ?? []),
-        inputs: arr("inputs", prior?.inputs ?? []),
-        dataNeeds: str("dataNeeds", prior?.dataNeeds ?? ""),
-        acceptanceCriteria: arr("acceptanceCriteria", prior?.acceptanceCriteria ?? []),
-        ...(mergedFix ? { fixContext: mergedFix } : {}),
-      };
-      try {
-        await updateFeatureBrief(buildId, brief as import("@/lib/feature-build-types").FeatureBrief);
-        await updateBuildHappyPathState(userId, {
-          intake: {
-            constrainedGoal: brief.title || null,
-          },
-        }, buildId);
-      } catch (err) {
-        const msg = err instanceof Error ? err.message : "Update failed";
-        return { success: false, error: msg, message: `Could not update brief: ${msg}. The brief can only be updated during the Ideate phase. You are past that phase — proceed with your current phase instead.` };
-      }
-      return { success: true, entityId: buildId, message: `Updated Feature Brief for ${buildId}` };
-    }
-
-    case "register_digital_product_from_build": {
-      const buildId = await resolveActiveBuildId(userId, extractBuildIdHint(params));
-      if (!buildId) return { success: false, error: "No active build", message: "No active build found" };
-      // Pre-flight: deploy_feature must have run first to extract the diff.
-      const diffCheck = await prisma.featureBuild.findUnique({
-        where: { buildId },
-        select: { diffPatch: true },
-      });
-      if (!diffCheck?.diffPatch) {
-        return {
-          success: false,
-          error: "deploy_feature must be called first",
-          message: "The sandbox diff has not been extracted yet. Call deploy_feature first to extract the diff, then call register_digital_product_from_build.",
-        };
-      }
-      const { shipBuild } = await import("@/lib/actions/build");
-      try {
-        const result = await shipBuild({
-          buildId,
-          name: String(params["name"]),
-          portfolioSlug: String(params["portfolioSlug"]),
-          versionBump: (params["versionBump"] as "major" | "minor" | "patch") ?? "minor",
-          // Thread the MCP actor through so this works in a session-less context
-          // (autonomous ship from the reconciler). UI callers go through the
-          // session as before; shipBuild falls back to requireBuildAccess().
-          actorUserId: userId,
-        });
-        return {
-          success: true,
-          entityId: result.productId,
-          message: result.message,
-          data: {
-            productInternalId: result.productInternalId,
-            portfolioInternalId: result.portfolioInternalId,
-            promotionId: result.promotionId,
-          },
-        };
-      } catch (err) {
-        const msg = err instanceof Error ? err.message : "Ship failed";
-        return { success: false, error: msg, message: `Product registration failed: ${msg}` };
-      }
-    }
-
-    case "create_build_epic": {
-      const epicBuildId = await resolveActiveBuildId(userId, extractBuildIdHint(params));
-      if (!epicBuildId) return { success: false, error: "No active build", message: "No active build found" };
-      // Auto-resolve digitalProductId and portfolioSlug from the build's linked product
-      const epicBuild = await prisma.featureBuild.findUnique({
-        where: { buildId: epicBuildId },
-        select: {
-          digitalProductId: true,
-          portfolioId: true,
-          digitalProduct: { select: { portfolio: { select: { slug: true } } } },
-        },
-      });
-      const resolvedProductId = epicBuild?.digitalProductId ?? undefined;
-      const resolvedPortfolioSlug = typeof params["portfolioSlug"] === "string"
-        ? params["portfolioSlug"]
-        : epicBuild?.digitalProduct?.portfolio?.slug ?? undefined;
-
-      const { createBuildEpic } = await import("@/lib/actions/build");
-      const epicInput: { buildId: string; title: string; portfolioSlug?: string; digitalProductId?: string } = {
-        buildId: epicBuildId,
-        title: String(params["title"]),
-      };
-      if (resolvedPortfolioSlug) epicInput.portfolioSlug = resolvedPortfolioSlug;
-      if (resolvedProductId) epicInput.digitalProductId = resolvedProductId;
-      try {
-        const result = await createBuildEpic(epicInput);
-        await updateBuildHappyPathState(userId, {
-          intake: {
-            epicId: result.epicId,
-          },
-        }, epicBuildId);
-        return { success: true, entityId: result.epicId, message: result.message };
-      } catch (err) {
-        const msg = err instanceof Error ? err.message : "Epic creation failed";
-        return { success: false, error: msg, message: `Could not create epic: ${msg}` };
-      }
-    }
-
     case "save_build_notes": {
       const buildId = await resolveActiveBuildId(userId, extractBuildIdHint(params));
       if (!buildId) return { success: false, error: "No active build", message: "No active build found" };
@@ -2462,34 +2140,6 @@ export async function executeTool(
         entityId: buildId,
         message: `Build progress visibility loaded for ${buildId}: ${projection.progress.primary.completed}/${projection.progress.primary.total} tasks from ${projection.progress.primary.source}.`,
         data: projection as unknown as Record<string, unknown>,
-      };
-    }
-
-    case "verification_preflight": {
-      const buildId = await resolveActiveBuildId(userId, extractBuildIdHint(params));
-      if (!buildId) return { success: false, error: "No active build", message: "No active build found" };
-      const build = await prisma.featureBuild.findUnique({
-        where: { buildId },
-        select: { phase: true, acceptanceMet: true, verificationOut: true, buildExecState: true },
-      });
-      if (!build) return { success: false, error: "Build not found", message: `Build ${buildId} was not found.` };
-      const { verificationPreflight, gatherPreflightSignals, preflightDirective } = await import(
-        "@/lib/build/verification-preflight"
-      );
-      // The portal serving this tool is up (installHealthy) and its DB is reachable
-      // (this row just loaded). Sandbox/quiescence probes are a follow-up refinement;
-      // for now they default healthy so the verdict turns on evidence + artifact.
-      const signals = gatherPreflightSignals(build, {
-        installHealthy: true,
-        requiredServicesHealthy: true,
-        explicitBlocker: null,
-      });
-      const result = verificationPreflight(signals);
-      return {
-        success: true,
-        entityId: buildId,
-        message: preflightDirective(result),
-        data: { verdict: result.verdict, reason: result.reason, blocker: result.blocker },
       };
     }
 
@@ -3932,52 +3582,6 @@ export async function executeTool(
       }
     }
 
-    case "start_build": {
-      const buildId = await resolveActiveBuildId(userId, extractBuildIdHint(params));
-      if (!buildId) return { success: false, error: "No active build.", message: "No active build." };
-
-      try {
-        const { assertFeatureBuildDependenciesSatisfied } = await import("@/lib/build/feature-build-dependencies");
-        await assertFeatureBuildDependenciesSatisfied({ db: prisma, buildId });
-      } catch (err) {
-        return {
-          success: false,
-          error: "Dependency gate blocked.",
-          message: getErrorMessage(err),
-        };
-      }
-
-      const { isSandboxAvailable, startBuildBranch } = await import("@/lib/integrate/sandbox/build-branch");
-
-      const available = await isSandboxAvailable();
-      if (!available) {
-        return {
-          success: false,
-          error: "Sandbox container is not running.",
-          message: "The sandbox is not running. Call check_sandbox to see the status, then start_sandbox if it is stopped.",
-        };
-      }
-
-      await startBuildBranch(buildId);
-
-      try {
-        const { startSandboxDevServer } = await import("@/lib/sandbox");
-        await startSandboxDevServer(process.env.SANDBOX_CONTAINER_ID ?? "dpf-sandbox-1");
-      } catch (devErr) {
-        console.log(`[start_build] preview server start failed (non-fatal): ${(devErr as Error).message?.slice(0, 100)}`);
-      }
-
-      const { agentEventBus } = await import("@/lib/agent-event-bus");
-      if (context?.threadId) agentEventBus.emit(context.threadId, { type: "phase:change", buildId, phase: "build" });
-      logBuildActivity(buildId, "start_build", `Build branch ready for ${buildId}.`);
-      return {
-        success: true,
-        message: `Build workspace ready. Sandbox is running on port ${process.env.SANDBOX_PORT ?? "3035"}. Start writing files.`,
-        entityId: buildId,
-        data: { containerId: process.env.SANDBOX_CONTAINER_ID ?? "dpf-sandbox-1", port: Number(process.env.SANDBOX_PORT ?? "3035") },
-      };
-    }
-
     case "generate_code":
     case "iterate_sandbox":
       // Removed: these tools caused runaway loops by spawning nested LLM calls.
@@ -4486,662 +4090,6 @@ export async function executeTool(
       return { success: false, error: "Unknown sandbox tool", message: "Internal error." };
     }
 
-    case "validate_schema": {
-      const buildId = await resolveActiveBuildId(userId, extractBuildIdHint(params));
-      if (!buildId) return { success: false, error: "No active build.", message: "No active build." };
-
-      // Simple availability check — no slot management
-      const { isSandboxAvailable } = await import("@/lib/integrate/sandbox/build-branch");
-      const vsAvailable = await isSandboxAvailable();
-      if (!vsAvailable) {
-        return {
-          success: false,
-          error: "Sandbox container is not running.",
-          message: "The sandbox (dpf-sandbox-1) is not running. Call start_build first.",
-        };
-      }
-      const vsSandboxId = process.env.SANDBOX_CONTAINER_ID ?? "dpf-sandbox-1";
-
-      try {
-        const { execInSandbox } = await import("@/lib/sandbox");
-        const schemaContent = await execInSandbox(
-          vsSandboxId,
-          "cat /workspace/packages/db/prisma/schema.prisma",
-        );
-        const { validatePrismaSchema, formatSchemaValidation } = await import("@/lib/integrate/schema-validator");
-        const result = validatePrismaSchema(schemaContent);
-
-        logBuildActivity(buildId, "validate_schema", result.summary);
-
-        if (!result.valid) {
-          return {
-            success: false,
-            error: "Schema validation failed",
-            message: formatSchemaValidation(result),
-            data: result as unknown as Record<string, unknown>,
-          };
-        }
-
-        return {
-          success: true,
-          message: formatSchemaValidation(result),
-          data: result as unknown as Record<string, unknown>,
-        };
-      } catch (err) {
-        return { success: false, error: "Schema validation error", message: err instanceof Error ? err.message : "Failed to validate schema" };
-      }
-    }
-
-    case "deploy_feature": {
-      const buildId = await resolveActiveBuildId(userId, extractBuildIdHint(params));
-      if (!buildId) return { success: false, error: "No active build.", message: "No active build." };
-      const build = await prisma.featureBuild.findUnique({
-        where: { buildId },
-        select: { sandboxId: true, buildBranch: true, phase: true, createdById: true, designDoc: true },
-      });
-      if (!build || build.createdById !== userId) {
-        return { success: false, error: "Build not found.", message: `No active build ${buildId} was found for this user.` };
-      }
-      if (!build?.sandboxId) return { success: false, error: "Sandbox not running.", message: "No sandbox." };
-      // Invariant: buildBranch must be set once we're in build phase or later.
-      // Its absence means startBuildBranch never ran — so the sandbox tree
-      // is on whatever HEAD happened to be (client branch or baseline), and
-      // any diff we extract will pick up leaked work from earlier builds.
-      // Refuse to deploy until start_build runs and registers a buildBranch.
-      if (!build.buildBranch) {
-        return {
-          success: false,
-          error: "Build branch not initialized.",
-          message: "This build has no buildBranch on record — start_build never completed. Run start_build to create and register build/<buildId>, then retry deploy_feature. Deploying without a build branch would include leaked changes from prior builds.",
-        };
-      }
-
-      const { diagnoseSandboxReadiness } = await import("@/lib/integrate/sandbox/sandbox-admin");
-      const { assertSandboxReadyForDeploy } = await import("@/lib/integrate/sandbox/sandbox-readiness-gate");
-      const readiness = await diagnoseSandboxReadiness({ buildId });
-      const readinessGate = assertSandboxReadyForDeploy(readiness);
-      if (!readinessGate.ok) {
-        logBuildActivity(buildId, "deploy_feature", readinessGate.message);
-        return {
-          success: false,
-          error: "Sandbox readiness blocked deploy_feature.",
-          message: readinessGate.message,
-          data: { ...readiness },
-        };
-      }
-
-      const devConfig = await prisma.platformDevConfig.findUnique({
-        where: { id: "singleton" },
-        select: { contributionMode: true, gitRemoteUrl: true },
-      });
-      const { getPlatformDevPolicyState } = await import("@/lib/platform-dev-policy");
-      const policyState = getPlatformDevPolicyState(devConfig);
-      if (policyState === "policy_pending") {
-        return {
-          success: false,
-          error: "Platform development policy not configured.",
-          message:
-            "Build Studio can keep editing and validating in the shared workspace, but production promotion stays blocked until Platform Development is configured in the portal. Go to Admin > Platform Development and choose whether this install stays private or can contribute upstream.",
-        };
-      }
-
-      // Extract diff from sandbox. Pass clientBranch as the diff base so
-      // committed work on the build branch (`git commit` from generate_code)
-      // is captured — without the base ref, `git diff --cached` only sees
-      // staged-but-uncommitted changes and returns empty for any build whose
-      // agent committed before deploy_feature ran.
-      const { extractAndCategorizeDiff, scanForDestructiveOps, isNowInWindow } = await import("@/lib/integrate/sandbox/sandbox-promotion");
-      const { getClientIdentity } = await import("@/lib/integrate/sandbox/build-branch");
-      const { clientBranch } = await getClientIdentity();
-      const extracted = await extractAndCategorizeDiff(build.sandboxId, { baseRef: clientBranch });
-      if (!extracted.fullDiff.trim()) {
-        await prisma.featureBuild.update({
-          where: { buildId },
-          data: { diffPatch: "", diffSummary: "" },
-        });
-        const noDiffMessage = "No releasable source changes were found in the sandbox. This build currently has only generated/cache churn or no real code changes, so release preparation cannot continue until implementation produces a real source diff.";
-        logBuildActivity(buildId, "deploy_feature", noDiffMessage);
-        return {
-          success: false,
-          error: "No releasable source changes found.",
-          message: noDiffMessage,
-          data: {
-            diffLength: 0,
-            codeFiles: 0,
-            migrationFiles: 0,
-          },
-        };
-      }
-
-      // Guard: schema regression check.
-      // If the diff removes existing fields/models from schema.prisma, the sandbox
-      // was initialized from a stale portal image and this diff would silently
-      // regress main's schema. Block deploy and surface the removed lines so the
-      // operator knows what drifted.
-      if (extracted.schemaRegressions.length > 0) {
-        const regressionSample = extracted.schemaRegressions.slice(0, 10).join("\n");
-        const regressionMessage =
-          `Schema regression detected in sandbox diff — ${extracted.schemaRegressions.length} existing field(s) or model declaration(s) would be removed from packages/db/prisma/schema.prisma. ` +
-          `This almost always means the sandbox was initialized from a portal image that predates recent schema changes on main. ` +
-          `Rebuild the sandbox from a fresh image (Admin → Build Studio → Rebuild Sandbox) and re-run the build before deploying.\n\n` +
-          `Removed lines (first 10):\n${regressionSample}`;
-        logBuildActivity(buildId, "deploy_feature", regressionMessage);
-        return {
-          success: false,
-          error: "Schema regression detected.",
-          message: regressionMessage,
-          data: {
-            schemaRegressions: extracted.schemaRegressions,
-            regressionCount: extracted.schemaRegressions.length,
-          },
-        };
-      }
-
-      // Capture commit hashes alongside the diff so the contribution flow
-      // (which pushes build/<buildId> upstream) has the exact list of
-      // commits being submitted. Without this, FB.gitCommitHashes stays
-      // empty for committed-work builds and contribute_to_hive cannot
-      // attribute the PR's commits back to specific FBs.
-      const { listSandboxCommitsAheadOfBase } = await import("@/lib/integrate/sandbox/sandbox");
-      const commitHashes = await listSandboxCommitsAheadOfBase(build.sandboxId, clientBranch);
-
-      await prisma.featureBuild.update({
-        where: { buildId },
-        data: {
-          diffPatch: extracted.fullDiff,
-          diffSummary: extracted.fullDiff.slice(0, 500),
-          gitCommitHashes: commitHashes,
-        },
-      });
-
-      // Compute + cache the per-change disposition suggestion (EP-1A78BAE1) so
-      // the ship UI / coworker can prefill Keep vs Share. The human still makes
-      // the final call via set_change_disposition; this only pre-fills.
-      // Non-fatal — a failed suggestion leaves the fail-closed default.
-      try {
-        const { suggestDisposition } = await import("@/lib/integrate/disposition");
-        const { loadPrivatePathPatterns, compilePrivatePathMatcher, stripPrivatePathsFromDiff } =
-          await import("@/lib/integrate/private-paths");
-        const outboundEmpty = !stripPrivatePathsFromDiff(
-          extracted.fullDiff,
-          compilePrivatePathMatcher(await loadPrivatePathPatterns({ prisma })),
-        ).kept.trim();
-        let reusabilityScope: "one_off" | "parameterizable" | "already_generic" | null = null;
-        const dd = build.designDoc as Record<string, unknown> | null;
-        const scope = (dd?.reusabilityAnalysis as { scope?: string } | undefined)?.scope;
-        if (scope === "one_off" || scope === "parameterizable" || scope === "already_generic") {
-          reusabilityScope = scope;
-        }
-        let orgSpecificHits = 0;
-        try {
-          const { runSanitizationScan } = await import("@/lib/integrate/contribution-review");
-          const san = await runSanitizationScan(extracted.fullDiff);
-          orgSpecificHits = san.mustFixCount;
-        } catch { /* sanitization optional */ }
-        const suggestion = suggestDisposition({ reusabilityScope, orgSpecificHits, outboundEmpty });
-        await prisma.featureBuild.update({
-          where: { buildId },
-          data: {
-            dispositionSuggested: suggestion.suggested,
-            dispositionSuggestionReason: suggestion.reason,
-            dispositionSource: "suggested",
-          },
-        });
-        logBuildActivity(buildId, "deploy_feature", `disposition suggestion: ${suggestion.suggested} — ${suggestion.reason}`);
-      } catch (err) {
-        console.warn("[deploy_feature] disposition suggestion failed:", err);
-      }
-
-      // Scan migrations for destructive operations
-      let destructiveWarnings: string[] = [];
-      if (extracted.hasMigrations) {
-        destructiveWarnings = scanForDestructiveOps(extracted.fullDiff);
-      }
-
-      // Check deployment window availability
-      let windowStatus = "No business profile configured — deployment unrestricted.";
-      try {
-        const profile = await prisma.businessProfile.findFirst({
-          where: { isActive: true },
-          include: { deploymentWindows: true, blackoutPeriods: true },
-        });
-        if (profile) {
-          const now = new Date();
-          const activeBlackout = profile.blackoutPeriods.find(
-            (bp) => bp.startAt <= now && bp.endAt >= now,
-          );
-          if (activeBlackout) {
-            windowStatus = `Blackout active until ${activeBlackout.endAt.toISOString()}.`;
-          } else {
-            const matchingWindows = profile.deploymentWindows.filter(
-              (w) => w.allowedChangeTypes.includes("normal") && w.allowedRiskLevels.includes("low"),
-            );
-            if (matchingWindows.length > 0) {
-              windowStatus = isNowInWindow(matchingWindows)
-                ? "Deployment window is open now."
-                : `Not in a deployment window. Available: ${matchingWindows.map((w) => `${w.name}: ${w.startTime}-${w.endTime}`).join("; ")}`;
-            } else {
-              windowStatus = "No deployment windows configured — deployment unrestricted.";
-            }
-          }
-        }
-      } catch {
-        // Non-fatal — window check is advisory at this stage
-      }
-
-      // Run change impact analysis (EP-BUILD-HANDOFF-002 Phase 2b)
-      let impactReport: Awaited<ReturnType<typeof import("@/lib/change-impact").analyzeChangeImpact>> | null = null;
-      let impactSummary = "";
-      try {
-        const { analyzeChangeImpact, formatImpactForChat } = await import("@/lib/change-impact");
-        impactReport = await analyzeChangeImpact(extracted.fullDiff);
-        impactSummary = formatImpactForChat(impactReport);
-      } catch (err) {
-        console.warn("[deploy_feature] impact analysis failed:", err);
-      }
-
-      // Resolve approval authority (EP-BUILD-HANDOFF-002 Phase 2b)
-      let authorityInfo = "";
-      try {
-        const { resolveApprovalAuthority, isCurrentUserTheAuthority, formatAuthorityForChat } = await import("@/lib/approval-authority");
-        const riskLevel = impactReport?.riskLevel ?? "low";
-        const authority = await resolveApprovalAuthority("deployment", "normal", riskLevel, userId);
-        const isSelf = isCurrentUserTheAuthority(authority, userId);
-        authorityInfo = formatAuthorityForChat(authority, isSelf);
-      } catch (err) {
-        console.warn("[deploy_feature] authority resolution failed:", err);
-      }
-
-      // Contribution mode awareness (EP-BUILD-HANDOFF-002 Phase 2e extension)
-      let contributionModeInfo = "";
-      try {
-        const mode = devConfig?.contributionMode ?? "private";
-
-        if ((mode === "private" || mode === "fork_only") && !devConfig?.gitRemoteUrl) {
-          // Count untracked shipped features for escalating warning
-          const untrackedCount = await prisma.featureBuild.count({
-            where: { phase: "complete", gitCommitHashes: { isEmpty: true } },
-          });
-
-          if (untrackedCount >= 5) {
-            contributionModeInfo = `**Warning:** You have ${untrackedCount} custom features with no backup. This represents significant business value that could be lost in a container rebuild, Docker update, or system recovery. Setting up a git repository takes about 10 minutes and protects all your customizations. See Admin > Platform Development.`;
-          } else if (untrackedCount >= 2) {
-            contributionModeInfo = `**Note:** You now have ${untrackedCount} custom features deployed without version control. If your Docker containers are rebuilt, these changes could be lost. I'd recommend setting up a git repository -- see Admin > Platform Development.`;
-          } else if (untrackedCount >= 1) {
-            contributionModeInfo = "Note: since no git repository is configured, customizations exist only in your production container. You can set up a repository in Admin > Platform Development to protect your work.";
-          }
-        }
-      } catch (err) {
-        console.warn("[deploy_feature] contribution mode check failed:", err);
-      }
-
-      const messageParts = [
-        `Diff extracted: ${extracted.codeFiles.length} code file(s), ${extracted.migrationFiles.length} migration(s).`,
-        windowStatus,
-      ];
-      if (destructiveWarnings.length > 0) {
-        messageParts.push(`WARNING: ${destructiveWarnings.length} destructive operation(s) detected: ${destructiveWarnings.join("; ")}`);
-      }
-      if (impactSummary) {
-        messageParts.push("", impactSummary);
-      }
-      if (authorityInfo) {
-        messageParts.push("", authorityInfo);
-      }
-      if (contributionModeInfo) {
-        messageParts.push("", contributionModeInfo);
-      }
-
-      logBuildActivity(buildId, "deploy_feature", messageParts.join(" "));
-
-      return {
-        success: true,
-        message: messageParts.join("\n"),
-        data: {
-          diffLength: extracted.fullDiff.length,
-          summary: extracted.fullDiff.slice(0, 500),
-          codeFiles: extracted.codeFiles.length,
-          migrationFiles: extracted.migrationFiles.length,
-          destructiveWarnings,
-          windowStatus,
-          impactReport,
-        },
-      };
-    }
-
-    // ─── Portal PR Creation & Merge ────────────────────────────────────────
-
-    case "create_portal_pr": {
-      const buildId = await resolveActiveBuildId(userId, extractBuildIdHint(params));
-      if (!buildId) return { success: false, error: "No active build.", message: "No active build." };
-
-      const build = await prisma.featureBuild.findUnique({
-        where: { buildId },
-        select: {
-          id: true, title: true, diffPatch: true, buildBranch: true,
-          description: true, gitCommitHashes: true, updatedAt: true, buildExecState: true,
-          verificationOut: true, acceptanceMet: true, phase: true,
-          designDoc: true, buildPlan: true,
-          disposition: true, dispositionSuggestionReason: true,
-          productVersions: {
-            take: 1,
-            orderBy: { shippedAt: "desc" },
-            select: {
-              id: true,
-              promotions: { take: 1, orderBy: { createdAt: "desc" }, select: { promotionId: true, status: true } },
-            },
-          },
-        },
-      });
-      if (!build) return { success: false, error: "Build not found.", message: "Build not found." };
-
-      const diff = (build.diffPatch ?? "") as string;
-      if (!diff.trim()) return { success: false, error: "No diff available.", message: "Run deploy_feature first to extract the diff." };
-
-      const { buildSandboxStateFromRecord, assertSandboxReadyForPromotion, serializePlanDocument } = await import("@/lib/build/sandbox-state");
-      const sandboxState = buildSandboxStateFromRecord({
-        buildBranch: build.buildBranch,
-        gitCommitHashes: build.gitCommitHashes,
-        diffPatch: diff,
-        updatedAt: build.updatedAt,
-        planDocument: typeof build.buildPlan === "string" ? build.buildPlan : serializePlanDocument(build.buildPlan),
-        description: build.description,
-        buildExecState: build.buildExecState,
-      });
-      const promotionGate = assertSandboxReadyForPromotion(sandboxState);
-      if (!promotionGate.ok) {
-        logBuildActivity(buildId, "create_portal_pr", promotionGate.message);
-        return {
-          success: false,
-          error: "Sandbox promotion integrity blocked PR creation.",
-          message: `${promotionGate.message}\n\n${promotionGate.failures.join("\n")}`,
-          data: {
-            gate: {
-              ok: false,
-              failures: promotionGate.failures,
-            },
-            sandbox: promotionGate.state,
-          },
-        };
-      }
-
-      // Resolve the portal's own repo from git remote
-      let repoOwner: string | null = null;
-      let repoName: string | null = null;
-      try {
-        const { getRemoteUrl } = await import("@/lib/git-utils");
-        const remoteUrl = await getRemoteUrl();
-        if (remoteUrl) {
-          const match = remoteUrl.match(/github\.com[/:]([^/]+)\/([^/.]+)/);
-          if (match) { repoOwner = match[1]; repoName = match[2]; }
-        }
-      } catch { /* git may not be available */ }
-
-      // Fallback to upstream URL if no local git
-      if (!repoOwner) {
-        const devConfig = await prisma.platformDevConfig.findUnique({
-          where: { id: "singleton" },
-          select: { upstreamRemoteUrl: true },
-        });
-        const url = devConfig?.upstreamRemoteUrl ?? "https://github.com/OpenDigitalProductFactory/opendigitalproductfactory.git";
-        const match = url.match(/github\.com[/:]([^/]+)\/([^/.]+)/);
-        if (match) { repoOwner = match[1]; repoName = match[2]; }
-      }
-
-      if (!repoOwner || !repoName) {
-        return { success: false, error: "Cannot determine repository.", message: "No git remote or upstream URL configured." };
-      }
-
-      // Public-egress boundary (Private/Public Change Segregation): the
-      // private/public filter applies ONLY when shipping to the PUBLIC hive. A
-      // PR to the install's OWN repo is its private home and keeps the full
-      // diff (proprietary paths included). This also closes the silent
-      // fall-back-to-public leak — a customer's proprietary build never goes to
-      // the canonical upstream unless that IS the configured target. Spec:
-      // docs/superpowers/specs/2026-06-19-hive-contribution-architecture-and-egress-model.md
-      let shareableDiff = diff;
-      {
-        const _egCfg = await prisma.platformDevConfig.findUnique({
-          where: { id: "singleton" },
-          select: { upstreamRemoteUrl: true },
-        });
-        const { classifyEgress } = await import("@/lib/integrate/contribution-egress");
-        const egress = classifyEgress({ owner: repoOwner, repo: repoName }, _egCfg?.upstreamRemoteUrl);
-        if (egress === "public-hive") {
-          // Fail-closed disposition gate (EP-1A78BAE1): a public-hive PR may only
-          // carry a change confirmed "shareable". Own-repo PRs skip this — that
-          // is the install's private home.
-          const { mayShareToPublicHive, privateDispositionBlockMessage } = await import("@/lib/integrate/disposition");
-          if (!mayShareToPublicHive(build.disposition)) {
-            logBuildActivity(buildId, "create_portal_pr", "blocked: change disposition is private (public-hive target)");
-            return {
-              success: false,
-              error: "Change is kept private.",
-              message: privateDispositionBlockMessage(build.dispositionSuggestionReason),
-            };
-          }
-          const { loadPrivatePathPatterns, compilePrivatePathMatcher, stripPrivatePathsFromDiff } =
-            await import("@/lib/integrate/private-paths");
-          shareableDiff = stripPrivatePathsFromDiff(
-            diff,
-            compilePrivatePathMatcher(await loadPrivatePathPatterns({ prisma })),
-          ).kept;
-          if (!shareableDiff.trim()) {
-            return {
-              success: false,
-              error: "Only private paths.",
-              message:
-                "This change only affects parts of your system you've marked private (see .dpf/private-paths or Admin > Platform Development), so there is nothing to share upstream.",
-            };
-          }
-        }
-      }
-
-      // Resolve token
-      const { resolveHiveToken, getPlatformIdentity, generatePrivateBranchName, generateAnonymousCommitMessage } = await import("@/lib/integrate/identity-privacy");
-      const token = await resolveHiveToken();
-      if (!token) {
-        return { success: false, error: "No GitHub token available.", message: "Configure HIVE_CONTRIBUTION_TOKEN or a git credential to create PRs." };
-      }
-
-      // Run pre-PR gates
-      const { runPrePRGates, formatGateReport } = await import("@/lib/integrate/pre-pr-gates");
-      const gateResult = runPrePRGates(shareableDiff);
-
-      // If gates block, return the report without creating a PR
-      if (!gateResult.canProceed) {
-        logBuildActivity(buildId, "create_portal_pr", `BLOCKED: ${gateResult.summary}`);
-        return {
-          success: false,
-          error: "Pre-PR gates failed.",
-          message: `The pre-PR security gates found blocking issues. Fix these before creating a PR.\n\n${formatGateReport(gateResult)}`,
-          data: { gates: gateResult },
-        };
-      }
-
-      // Build the PR
-      const platformId = await getPlatformIdentity();
-      const branchName = generatePrivateBranchName(platformId.clientId, build.title);
-
-      const devConfigForDco = await prisma.platformDevConfig.findUnique({
-        where: { id: "singleton" },
-        select: { dcoAcceptedAt: true },
-      });
-
-      const commitMessage = generateAnonymousCommitMessage({
-        title: build.title,
-        buildId,
-        productId: null,
-        platformIdentity: platformId,
-        dcoAcceptedAt: devConfigForDco?.dcoAcceptedAt ?? undefined,
-      });
-
-      // Build PR body with gate report and build evidence
-      const verification = build.verificationOut as Record<string, unknown> | null;
-      const typecheckPassed = verification?.typecheckPassed === true;
-      const testsPassed = typeof verification?.testsPassed === "number" ? verification.testsPassed : 0;
-      const testsFailed = typeof verification?.testsFailed === "number" ? verification.testsFailed : 0;
-
-      // acceptanceMet (Json?) stores either a bare array or {acceptanceCriteria: [...]} — `.filter` must not assume array shape.
-      const rawAcceptance = build.acceptanceMet as unknown;
-      const acceptance: Array<{ met?: boolean }> = Array.isArray(rawAcceptance)
-        ? (rawAcceptance as Array<{ met?: boolean }>)
-        : rawAcceptance && typeof rawAcceptance === "object" && Array.isArray((rawAcceptance as { acceptanceCriteria?: unknown }).acceptanceCriteria)
-          ? ((rawAcceptance as { acceptanceCriteria: Array<{ met?: boolean }> }).acceptanceCriteria)
-          : [];
-      const acMet = acceptance.filter((a) => a?.met === true).length;
-      const acTotal = acceptance.length;
-
-      const prBody = [
-        `## ${build.title}`,
-        "",
-        `Build: \`${buildId}\` | Phase: \`${build.phase}\``,
-        "",
-        "### Verification",
-        `- TypeCheck: ${typecheckPassed ? "PASSED" : "FAILED"}`,
-        `- Tests: ${testsPassed} passed, ${testsFailed} failed`,
-        `- Acceptance: ${acMet}/${acTotal} criteria met`,
-        "",
-        formatGateReport(gateResult),
-        "",
-        "---",
-        `License: Apache-2.0 | ${platformId.dcoSignoff}`,
-      ].join("\n");
-
-      const prTitle = `feat(${buildId}): ${build.title}`;
-      const labels = ["build-studio", "automated"];
-      if (gateResult.requiresHumanReview) labels.push("needs-review");
-      if (!typecheckPassed || testsFailed > 0) labels.push("verification-issues");
-
-      const { createBranchAndPR, mergePR, commentOnPR } = await import("@/lib/integrate/github-api-commit");
-
-      const prResult = await createBranchAndPR({
-        headOwner: repoOwner,
-        headRepo: repoName,
-        baseOwner: repoOwner,
-        baseRepo: repoName,
-        branchName,
-        commitMessage,
-        diff: shareableDiff,
-        prTitle,
-        prBody,
-        labels,
-        token,
-      });
-
-      if (!prResult.prUrl || !prResult.prNumber) {
-        logBuildActivity(buildId, "create_portal_pr", `Branch created (${branchName}) but PR creation failed.`);
-        return {
-          success: false,
-          error: "PR creation failed.",
-          message: `Branch \`${branchName}\` was created with the commit, but the pull request could not be opened. Check GitHub permissions.`,
-          data: { branchName, commitSha: prResult.commitSha },
-        };
-      }
-
-      // Capture the PR onto the build's Work Capsule so it becomes a queryable,
-      // surfaceable fact (delivery visibility). The capsule's pullRequestUrl/
-      // Number are READ by the change-lanes projection, the runtime-target
-      // rollup, and the capsule presenters, but had no writer at PR-creation
-      // time — so a build's PR stayed invisible until a delayed branch-name-
-      // matched GitHub inventory snapshot backfilled it (or never). Migration-
-      // free + best-effort: opening the PR must never hinge on this write.
-      try {
-        const { captureBuildPrOntoCapsule } = await import("@/lib/build/capture-build-pr");
-        const cap = await captureBuildPrOntoCapsule({
-          db: prisma,
-          featureBuildId: build.id,
-          prNumber: prResult.prNumber,
-          prUrl: prResult.prUrl,
-        });
-        if (cap.captured > 0) {
-          logBuildActivity(buildId, "create_portal_pr", `Linked PR #${prResult.prNumber} to ${cap.captured} work capsule(s).`);
-        }
-      } catch (err) {
-        console.warn("[create_portal_pr] PR-capture onto capsule failed:", err);
-      }
-
-      // Auto-merge decision: all gates pass + build fully verified
-      const fullyVerified = typecheckPassed && testsFailed === 0 && acMet === acTotal && acTotal > 0;
-      let merged = false;
-
-      if (!gateResult.requiresHumanReview && fullyVerified) {
-        // Auto-merge via squash
-        const mergeResult = await mergePR({
-          owner: repoOwner,
-          repo: repoName,
-          prNumber: prResult.prNumber,
-          commitTitle: `${prTitle} (#${prResult.prNumber})`,
-          mergeMethod: "squash",
-          token,
-        });
-        merged = mergeResult.merged;
-
-        if (merged) {
-          // Update lifecycle: FeatureBuild → complete, ChangePromotion → deployed
-          await prisma.featureBuild.update({
-            where: { buildId },
-            data: { phase: "complete" },
-          });
-          const { recordReadyDependentsAfterCompletion } = await import("@/lib/build/feature-build-dependencies");
-          await recordReadyDependentsAfterCompletion({ db: prisma, buildId }).catch((err) => {
-            console.warn("[create_portal_pr] dependency readiness check failed:", err);
-          });
-
-          const promotion = build.productVersions[0]?.promotions[0];
-          if (promotion) {
-            await prisma.changePromotion.update({
-              where: { promotionId: promotion.promotionId },
-              data: { status: "deployed", deployedAt: new Date(), deploymentLog: `Auto-merged PR #${prResult.prNumber}` },
-            });
-          }
-
-          logBuildActivity(buildId, "create_portal_pr", `PR #${prResult.prNumber} auto-merged. Build complete.`);
-        } else {
-          // Merge failed — post comment explaining why
-          await commentOnPR({
-            owner: repoOwner, repo: repoName, prNumber: prResult.prNumber,
-            body: `Auto-merge failed. This PR requires manual review and merge.\n\n${formatGateReport(gateResult)}`,
-            token,
-          }).catch(() => {});
-
-          logBuildActivity(buildId, "create_portal_pr", `PR #${prResult.prNumber} created but auto-merge failed.`);
-        }
-      } else {
-        // Needs human review — post gate report as comment
-        const reasons: string[] = [];
-        if (gateResult.requiresHumanReview) reasons.push("security gate warnings");
-        if (!typecheckPassed) reasons.push("TypeCheck failed");
-        if (testsFailed > 0) reasons.push(`${testsFailed} test(s) failed`);
-        if (acMet < acTotal) reasons.push(`${acTotal - acMet} acceptance criteria not met`);
-
-        await commentOnPR({
-          owner: repoOwner, repo: repoName, prNumber: prResult.prNumber,
-          body: `This PR requires human review: ${reasons.join(", ")}.\n\n${formatGateReport(gateResult)}`,
-          token,
-        }).catch(() => {});
-
-        logBuildActivity(buildId, "create_portal_pr", `PR #${prResult.prNumber} created — needs review: ${reasons.join(", ")}`);
-      }
-
-      const statusMsg = merged
-        ? `PR #${prResult.prNumber} created and auto-merged. Build is complete.`
-        : `PR #${prResult.prNumber} created and awaiting review. ${gateResult.summary}`;
-
-      return {
-        success: true,
-        message: `${statusMsg}\n\n${prResult.prUrl}`,
-        data: {
-          prUrl: prResult.prUrl,
-          prNumber: prResult.prNumber,
-          branchName,
-          commitSha: prResult.commitSha,
-          merged,
-          gates: gateResult,
-        },
-      };
-    }
-
     // ─── Hive Mind Contribution ──────────────────────────────────────────────
 
     case "assess_contribution": {
@@ -5272,33 +4220,6 @@ export async function executeTool(
       logBuildActivity(buildId, "assess_contribution", `Recommendation: ${recommendation}. Vision: ${visionScore}, Community: ${communityScore}, Type: ${augLevel}, Sensitive: ${isSensitive}`);
 
       return { success: true, message: assessment.summary, data: assessment };
-    }
-
-    case "set_change_disposition": {
-      const buildId = await resolveActiveBuildId(userId, extractBuildIdHint(params));
-      if (!buildId) return { success: false, error: "No active build.", message: "No active build." };
-      const disposition = (params.disposition as string) ?? "";
-      if (disposition !== "private" && disposition !== "shareable") {
-        return { success: false, error: "Invalid disposition.", message: "disposition must be 'private' or 'shareable'." };
-      }
-      const reason = typeof params.reason === "string" ? params.reason : null;
-      await prisma.featureBuild.update({
-        where: { buildId },
-        data: {
-          disposition,
-          dispositionReason: reason,
-          dispositionSource: "operator",
-          dispositionDecidedAt: new Date(),
-          dispositionDecidedById: userId,
-        },
-      });
-      logBuildActivity(buildId, "set_change_disposition", `disposition=${disposition}${reason ? ` (${reason})` : ""}`);
-      return {
-        success: true,
-        message: disposition === "shareable"
-          ? "This change is marked to share with the community. Run contribute_to_hive (or create the portal PR) to share it."
-          : "This change is kept on your system and will not be shared.",
-      };
     }
 
     case "contribute_to_hive": {
@@ -5671,221 +4592,6 @@ export async function executeTool(
         success: true,
         message: `Feature Pack ${packId} created and contributed to the Hive Mind. ${manifest.totalFiles} file(s) packaged with DCO attestation.${prMessage} Thank you for contributing!`,
         data: { packId, manifest, dcoAttestation, prUrl },
-      };
-    }
-
-    case "run_ux_test": {
-      const buildId = await resolveActiveBuildId(userId, extractBuildIdHint(params));
-      if (!buildId) return { success: false, error: "No active build.", message: "No active build." };
-      const build = await prisma.featureBuild.findUnique({ where: { buildId }, select: { sandboxId: true, sandboxPort: true, brief: true, kind: true } });
-      if (!build?.sandboxPort || !build.sandboxId || !build.brief) return { success: false, error: "Sandbox or brief not ready.", message: "Launch sandbox and save brief first." };
-
-      const { deriveFixUxTestCases } = await import("@/lib/explore/feature-build-types");
-      const brief = build.brief as {
-        acceptanceCriteria?: string[];
-        fixContext?: import("@/lib/explore/feature-build-types").FixContext;
-      };
-      // Explicit `tests` always win. Otherwise, for a fix build derive the
-      // assertion from the structured fix diagnosis (defect-gone on its route)
-      // rather than the polluted feature acceptanceCriteria. (BI-AC5CFDB0)
-      const testCases =
-        (params.tests as string[] | undefined) ??
-        (build.kind === "fix"
-          ? deriveFixUxTestCases(brief.fixContext)
-          : brief.acceptanceCriteria ?? []);
-      if (testCases.length === 0) return { success: false, error: "No test cases.", message: build.kind === "fix" ? "No fix context (route/expected) or test cases to verify." : "No acceptance criteria or test cases to run." };
-
-      try {
-        const BROWSER_USE_URL = process.env.BROWSER_USE_URL || "http://browser-use:8500/mcp";
-        // browser-use runs inside the docker compose network — use the
-        // internal service URL (http://sandbox:3000), not the host port,
-        // so assets and API calls resolve correctly.
-        const { resolveSandboxUrl } = await import("@/lib/integrate/sandbox/resolve-sandbox-url");
-        const sandboxUrl = resolveSandboxUrl(build.sandboxId, build.sandboxPort).internal;
-
-        const testRes = await fetch(BROWSER_USE_URL, {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({
-            jsonrpc: "2.0",
-            id: 1,
-            method: "tools/call",
-            params: {
-              name: "browse_run_tests",
-              arguments: {
-                url: sandboxUrl,
-                tests: testCases,
-                // Scope per-step screenshots to a build-specific subdir on
-                // the shared /evidence volume. Portal serves them through
-                // /api/build/<buildId>/evidence/<file>.png.
-                evidence_dir: `build_${buildId}`,
-              },
-            },
-          }),
-          signal: AbortSignal.timeout(300000), // 5 min for full test suite
-        });
-        const testResult = await testRes.json();
-        const testContent = JSON.parse(testResult?.result?.content?.[0]?.text ?? "{}");
-
-        // Convert to UxTestStep format for storage. screenshot_path (when
-        // present) is a filename inside the evidence_dir — turn it into a
-        // portal-served URL the ReviewPanel can render.
-        const steps = (testContent.results ?? []).map((r: Record<string, unknown>, i: number) => ({
-          step: (r.test as string) ?? `Test ${i + 1}`,
-          passed: r.status === "pass",
-          screenshotUrl: typeof r.screenshot_path === "string"
-            ? `/api/build/${encodeURIComponent(buildId)}/evidence/${encodeURIComponent(r.screenshot_path)}`
-            : null,
-          error: r.status !== "pass" ? ((r.detail as string) ?? null) : null,
-        }));
-
-        const { agentEventBus } = await import("@/lib/agent-event-bus");
-        for (let i = 0; i < steps.length; i++) {
-          if (context?.threadId) {
-            agentEventBus.emit(context.threadId, {
-              type: "test:step",
-              stepIndex: i,
-              description: steps[i]!.step,
-              passed: steps[i]!.passed,
-            });
-          }
-        }
-        const { saveBuildArtifactRevision } = await import("@/lib/build/build-artifact-provenance");
-        await saveBuildArtifactRevision({
-          buildId,
-          field: "uxTestResults",
-          savedByAgentId: context?.agentId ?? null,
-          savedByUserId: userId,
-          threadId: context?.threadId ?? null,
-          value: steps,
-        });
-        if (context?.threadId) agentEventBus.emit(context.threadId, { type: "evidence:update", buildId, field: "uxTestResults" });
-        const passed = steps.filter((s: { passed: boolean }) => s.passed).length;
-        logBuildActivity(buildId, "run_ux_test", `UX tests: ${passed}/${steps.length} passed (browser-use).`);
-        return {
-          success: true,
-          message: `UX tests: ${passed}/${steps.length} passed.`,
-          data: { buildId, steps, browserUseResults: testContent },
-        };
-      } catch (err) {
-        const msg = (err as Error).message?.slice(0, 200) ?? "Unknown error";
-        return { success: false, error: `UX test run failed: ${msg}`, message: `UX verification service (browser-use) is unreachable. Run 'docker compose up -d browser-use' or check the browser-use container logs. You can skip UX tests and proceed with the review if you have to.` };
-      }
-    }
-
-    case "start_ideate_research": {
-      // This tool is a signal — the actual research dispatch happens in
-      // agent-coworker.ts after the agentic loop returns. We just persist
-      // the user context so the dispatch knows what to research.
-      const scope = String(params.reusabilityScope ?? "parameterizable");
-      const userCtx = String(params.userContext ?? "");
-
-      // BI-F4A30FCB (Dale dogfood 2026-05-24): resolve the target build
-      // from agent context first. The previous "findFirst by phase=ideate
-      // ordered by updatedAt" silently mis-targeted whenever multiple
-      // builds were in ideate concurrently — the user's request landed on
-      // an unrelated build whose updatedAt happened to be newer.
-      // An explicit buildId param wins over the ambient conversation context:
-      // it lets an operator/agent drive a chosen build's research even when
-      // several builds are in-flight (autonomous batch). When omitted, the
-      // exact prior ambient-resolution behavior is preserved.
-      const activeBuild = await resolveIdeateBuildForTool({
-        contextBuildId: extractBuildIdHint(params) ?? context?.featureBuildId,
-        toolName: "start_ideate_research",
-      });
-      if (!activeBuild.build) {
-        return activeBuild.refusal;
-      }
-
-      await prisma.featureBuild.update({
-        where: { buildId: activeBuild.build.buildId },
-        data: {
-          buildExecState: {
-            ideateResearchRequested: true,
-            reusabilityScope: scope,
-            userContext: userCtx,
-            requestedAt: new Date().toISOString(),
-          },
-        },
-      });
-
-      return {
-        success: true,
-        message: "Research started. Searching the codebase and drafting the design document — this takes about 1-2 minutes. Tell the user you're researching now. IMPORTANT: Do NOT call saveBuildEvidence with field 'designDoc' — the research system saves the design document and runs the review automatically when research completes. Just wait and tell the user.",
-        data: { reusabilityScope: scope, userContext: userCtx, buildId: activeBuild.build.buildId },
-      };
-    }
-
-    case "start_scout_research": {
-      // Scout dispatch: similar to ideate research, but runs a fast parallel search + URL fetch
-      const externalUrls = (params.externalUrls as string[] | undefined) ?? [];
-
-      // BI-F4A30FCB (Dale dogfood 2026-05-24): resolve the target build
-      // from agent context first; see start_ideate_research comment above.
-      // BI-7FEAFD9A (2026-07-05): prefer the FB- hint from params.buildId
-      // (set by the executeTool preamble from the context cuid) — passing the
-      // raw context.featureBuildId cuid here made the resolver's
-      // where:{buildId} lookup a guaranteed miss, so scout always refused.
-      const resolved = await resolveIdeateBuildForTool({
-        contextBuildId: extractBuildIdHint(params) ?? context?.featureBuildId,
-        toolName: "start_scout_research",
-      });
-      if (!resolved.build) {
-        return resolved.refusal;
-      }
-      const activeBuild = await prisma.featureBuild.findUnique({
-        where: { buildId: resolved.build.buildId },
-        select: { buildId: true, buildExecState: true, scoutFindings: true },
-      });
-      if (!activeBuild) {
-        return { success: false, message: "Active ideate build vanished between resolve and read — try again." };
-      }
-
-      const current = activeBuild.buildExecState as Record<string, unknown> | null;
-
-      // Idempotency: if scout has already delivered findings, do NOT re-run.
-      // The agentic loop otherwise re-calls this tool every iteration because
-      // the initial response says "results will appear on the next turn" and
-      // the model doesn't see the findings in its prompt context. The stuck-
-      // detector eventually bails, but not before burning 4-5 iterations and
-      // preventing phase advance. Tell the agent plainly the work is done.
-      if (activeBuild.scoutFindings !== null && activeBuild.scoutFindings !== undefined) {
-        return {
-          success: true,
-          message:
-            "Scout already ran for this build. Findings are saved to Build Studio Context — proceed with ideate using the existing scout results; do NOT call start_scout_research again.",
-          data: { alreadyComplete: true },
-        };
-      }
-
-      // Idempotency: if a scout request is already pending dispatch (flag set
-      // by a prior call but not yet cleared by the coworker post-turn hook),
-      // don't stack up another request. Same guidance.
-      if (current?.scoutResearchRequested === true) {
-        return {
-          success: true,
-          message:
-            "Scout already requested and is running now. Wait for the next turn to see findings — do NOT call start_scout_research again.",
-          data: { alreadyRequested: true },
-        };
-      }
-
-      await prisma.featureBuild.update({
-        where: { buildId: activeBuild.buildId },
-        data: {
-          buildExecState: {
-            ...(current ?? {}),
-            scoutResearchRequested: true,
-            scoutUrls: externalUrls,
-            scoutRequestedAt: new Date().toISOString(),
-          },
-        },
-      });
-
-      return {
-        success: true,
-        message: "Scout started. Codebase search and URL parsing running in background — takes about 30 seconds. Results will appear in your Build Studio Context on the next turn. Do NOT call start_scout_research again; wait for the results.",
-        data: { urlCount: externalUrls.length },
       };
     }
