+++ apps/web/lib/mcp/pack-registry.ts
@@ -74,6 +74,8 @@ import { decompositionPack } from "./packs/decomposition-pack";
 
 import { grokSigninPack } from "./packs/grok-signin-pack";
 import { releasePack } from "./packs/release-pack";
+import { buildLifecyclePack } from "./packs/build-lifecycle-pack";
+import { buildChangePack } from "./packs/build-change-pack";
 
 export const TOOL_PACK_REGISTRY = composeToolPacks([
   deliberationSiemPack,
@@ -131,6 +133,8 @@ export const TOOL_PACK_REGISTRY = composeToolPacks([
   platformUtilitiesPack,
   backlogPack,
   releasePack,
+  buildLifecyclePack,
+  buildChangePack,
   adminPack,
   decompositionPack,
 ]);
