+++ scripts/mcp-tool-pack-baseline.json
@@ -2,9 +2,6 @@
   "assess_contribution",
   "check_sandbox",
   "contribute_to_hive",
-  "create_build_epic",
-  "create_portal_pr",
-  "deploy_feature",
   "diagnose_sandbox",
   "edit_sandbox_file",
   "generate_code",
@@ -30,28 +27,19 @@
   "record_functional_failure_evidence",
   "record_local_integration_result",
   "recover_sandbox",
-  "register_digital_product_from_build",
   "reviewBuildPlan",
   "reviewDesignDoc",
   "run_endpoint_tests",
   "run_sandbox_command",
   "run_sandbox_tests",
   "run_tool_script",
-  "run_ux_test",
   "saveBuildEvidence",
   "save_build_notes",
   "save_phase_handoff",
   "search_sandbox",
-  "set_change_disposition",
-  "start_build",
-  "start_ideate_research",
   "start_sandbox",
-  "start_scout_research",
   "submit_feedback",
-  "update_feature_brief",
   "update_lifecycle",
-  "validate_schema",
-  "verification_preflight",
   "verify_live_install_readiness",
   "write_sandbox_file"
 ]
